autosynth-graph: hash-interned value state graph IR
New crate: sea-of-nodes IR where every value is a VRegRef (u64 hash
handle into a pool of VRegStates). Operations take VRegRefs in and
produce VRegRefs out. Structural equality via hash gives free CSE.

- VRegState: width, preg, target, slot, dirty, const, op
- Op: opcode + SmallVec of input VRegRefs
- Pool: HashMap<u64, VRegState> intern pool
  - define_const, define_param, unary, binary, set_slot, assign_preg
- VRegRef: u64 hash handle — the universal operand type
- No separate Operand enum — VRegStates ARE operands

This replaces the flat VCode stream architecture. Migration is
incremental — old stream code stays untouched for now.

+++ crates/autosynth-ir/src/lib.rs
@@ -353,7 +353,7 @@ pub struct SlotRef {
     /// Base physical register (e.g. frame pointer).
     pub base: PReg,
     /// Byte offset from the base register.
-    pub offset: u32,
+    pub offset: u16,
 }
 
 /// Per-VReg state. Tracks where a value currently lives.
@@ -572,10 +572,10 @@ pub enum VCode {
     Load,
 
     /// Store value to memory location
+    /// - Operand::VReg -> value
     /// - Operand::PReg -> base
     /// - Operand::Const -> offset
     /// - VCode::Store
-    /// - Operand::DstPreg | Operand::DstVreg
     Store,
 
     /// Move / copy.
